entities directory structure
