+++ src/main/java/pl/lcc/yasmart/entity/project/NewCampaignDTO.java
@@ -1,4 +1,4 @@
-package pl.lcc.yasmart.entity;
+package pl.lcc.yasmart.entity.project;
 
 import pl.lcc.yasmart.common.reward.Reward;
