+++ src/main/java/pl/lcc/yasmart/entity/task/Task.java
@@ -1,20 +1,19 @@
-package pl.lcc.yasmart.entity;
+package pl.lcc.yasmart.entity.task;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.Accessors;
 import org.hibernate.annotations.NaturalId;
 import pl.lcc.yasmart.common.account.Account;
-import pl.lcc.yasmart.common.flow.ScenarioState;
+import pl.lcc.yasmart.common.flow.TaskState;
 import pl.lcc.yasmart.common.reward.Reward;
 import pl.lcc.yasmart.common.tag.Tag;
+import pl.lcc.yasmart.entity.sprint.Sprint;
 
 import java.time.LocalDateTime;
 import java.util.Set;
 import java.util.UUID;
-
 
 @Entity
 @Accessors(chain = true)
@@ -22,7 +21,7 @@ import java.util.UUID;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Sprint {
+public class Task {
 
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
@@ -38,7 +37,7 @@ public class Sprint {
     private LocalDateTime finished;
 
     @Enumerated(EnumType.STRING)
-    ScenarioState state;
+    private TaskState state;
 
     @ManyToMany
     private Set<Tag> tags;
@@ -48,23 +47,30 @@ public class Sprint {
 
     @JsonBackReference
     @ManyToOne(fetch = FetchType.LAZY)
-    private Campaign campaign;
-
-    @JsonManagedReference
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "sprint", cascade = CascadeType.ALL)
-    private Set<Task> tasks;
+    private Sprint sprint;
 
     @JsonBackReference
     @ManyToOne(fetch = FetchType.LAZY)
     private Account owner;
 
-    public Sprint addTask(Task task) {
-        task.setSprint(this);
-        tasks.add(task);
-        return this;
-    }
-    public Sprint addTag(Tag tag){
+    public Task addTag(Tag tag){
         tags.add(tag);
         return this;
+    }
+
+    @Override
+    public String toString() {
+        return "Task{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", description='" + description + '\'' +
+                ", created=" + created +
+                ", finished=" + finished +
+                ", state=" + state +
+               // ", tags=" + (tags!=null ? tags.stream().map(t -> t.getId()).toList() : null) +
+               // ", rewards=" + rewards +
+              //  ", sprint=" + sprint +
+                ", owner=" + (owner!= null ? owner.getName() : null) +
+                '}';
     }
 }
