+++ src/test/java/pl/lcc/yasmart/entity/CampaignControllerTest.java
@@ -8,6 +8,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 import pl.lcc.yasmart.common.loader.DataPreLoader;
+import pl.lcc.yasmart.entity.project.CampaignRepository;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
